requirements.txt
# Core RL and training requirements (without heavy optional dependencies)
gymnasium
stable-baselines3

# Deep Learning & Graphs
torch
torch-geometric
networkx

# Data & Math
pandas
numpy
scikit-learn
yfinance
ta

# Visualization (essential)
matplotlib
seaborn

# Dashboard
streamlit

# Config
pyyaml

# Explainability
shap
